Initial CPAOT support for large version bubble
This change modifies the existing --inputbubble CPAOT option
to work the same as Crossgen /largeversionbubble i.e. the main module
and all reference assemblies are considered to belong to the same
version bubble.

The gist of the change deals with encoding module overrides in
signatures. I have implemented a new R2R header table
ManifestMetadataTableNode and added logic for composing and emitting
the extra AssemblyRef list.

One challenging bit was that we need to emit all the signatures before
emitting the ManifestMetadataTableNode because only the emission
of the signatures shakes out all the required AssemblyRef's. I have
added a simple loop to ImportSectionsTableNode / ImportSectionNode
that traverses and emits all registered import cell signatures.

After I managed to make the basic signature emission work with module
override, I continued hitting issues with the two entrypoint tables.
For the (non-generic) method entrypoint table we need to filter out
methods outside of the current input module because the larger
compilation group representing the larger version bubble ends up
expanding the dependency graph to framework methods.

While these "external" methods are not compiled into the output
executable (except when they are occasionally inlined into a local
method), we were still trying to emit them into the entrypoint
table and we were crashing compilation because we couldn't find
some of the necessary tokens.

For instance entrypoints, I didn't find any manner of encoding
a module override for their signatures in the native hashtable so
for now I added filtering that we're only storing those instance
entrypoints that have the current input module as the reference
module. One can theoretically imagine that a module may contain
arbitrary generic instantiations spanning other modules when
larger bubbles are on.

While working on the change I also realized that the current way
of basing signature contexts on input modules in CorInfoImpl.ReadyToRun
was incorrect - when a method gets inlined, we still need the original
signature context because that's what defines the ambient module
used in the CoreCLR signature parser.

I have deleted the field and instead I added a new method
GetSignatureContext which currently returns
ReadyToRunCoregenNodeFactory.InputModuleContext. Once we start
implementing "single-file" (i.e. compiling multiple MSIL modules
into a single PE), we'll change this to something more fine-grained,
most likely based on some lookup on MethodBeingCompiled.

Thanks

Tomas